feat: buyer create review for each products in an order

+++ FoodConnect.Backend.Application/Commons/Constants/ShippingFeeConstant.cs
@@ -2,16 +2,16 @@
 {
     public static class ShippingFeeConstant
     {
-        public const double EXPRESS_BASE_FEE = 5000;
+        public const decimal EXPRESS_BASE_FEE = 5000;
         
-        public const double EXPRESS_PRICE_PER_KM = 5000;
+        public const decimal EXPRESS_PRICE_PER_KM = 5000;
         
-        public const double EXPRESS_FREE_DISTANCE = 2.0;
+        public const decimal EXPRESS_FREE_DISTANCE = 2.0m;
         
-        public const double EXPRESS_MAX_DISTANCE = 20.0;
+        public const decimal EXPRESS_MAX_DISTANCE = 20.0m;
 
-        public const double STANDARD_SAME_PROVINCE_FEE = 20000;
+        public const decimal STANDARD_SAME_PROVINCE_FEE = 20000;
         
-        public const double STANDARD_DIFFERENT_PROVINCE_FEE = 30000;
+        public const decimal STANDARD_DIFFERENT_PROVINCE_FEE = 30000;
     }
 }
